feat: issue #7  #3 -오라클 연결 기능 생성 - mode 설정 추가 - 모니터 크기 줄임
