--- .idea/libraries/ojdbc8.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="ojdbc8">
-    <CLASSES>
-      <root url="jar://$PROJECT_DIR$/ojdbc8.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
