+++ src/syncMonitor/Main.java
@@ -1,42 +1,34 @@
 package syncMonitor;
 
-import syncMonitor.config.DbConfig;
-import syncMonitor.config.MonitorConfig;
-import syncMonitor.session.SyncMonitorSession;
-import syncMonitor.session.SyncMonitorSessionOracle;
-import syncMonitor.session.SyncMonitorSessionTibero;
-import syncMonitor.view.ViewOracle;
-import syncMonitor.view.ViewTibero;
+import syncMonitor.config.wrapper.DbConfig.OracleConfig;
+import syncMonitor.config.wrapper.DbConfig.TiberoConfig;
+import syncMonitor.config.YmlConfigWrapper;
+import syncMonitor.config.wrapper.MonitorConfig;
+import syncMonitor.mode.BothTiberoOracle;
+import syncMonitor.mode.OnlyTibero;
+import syncMonitor.mode.view.View;
 
 public class Main {
 
     public static void main(String[] args) {
-        MonitorConfig config = new MonitorConfig();
-        DbConfig tiberoConfig = config.getTiberoConfig();
-        DbConfig oracleConfig = config.getOracleConfig();
+        YmlConfigWrapper config = new YmlConfigWrapper();
+        TiberoConfig tiberoConfig = config.getTiberoConfig();
+        OracleConfig oracleConfig = config.getOracleConfig();
+        MonitorConfig monitorConfig = config.getMonitorConfig();
+
         System.out.println("Tibero Configuration:");
         System.out.println(tiberoConfig);
 
         System.out.println("\nOracle Configuration:");
         System.out.println(oracleConfig);
 
-        // connection
-        SyncMonitorSession sessionTibero = SyncMonitorSessionTibero.getSyncMonitorSession(
-                tiberoConfig.getIp(),
-                tiberoConfig.getPort(),
-                tiberoConfig.getId(),
-                tiberoConfig.getPwd(),
-                tiberoConfig.getDbSid()
-        );
+        //BothTiberoOracle bothTiberoOracle = new BothTiberoOracle(oracleConfig, tiberoConfig);
+       // View viewBoth = bothTiberoOracle.getViewOracleTibero();
+        //viewBoth.doPrint();
 
-
-        ViewTibero viewTibero = new ViewTibero(sessionTibero,
-                tiberoConfig.getDbLink(),
-                tiberoConfig.getUser1(),
-                tiberoConfig.getUser2()
-        );
+        OnlyTibero onlyTibero = new OnlyTibero(tiberoConfig);
+        View viewTibero = onlyTibero.getViewTibero();
         viewTibero.doPrint();
-
     }
 
 }
