--- src/syncMonitor/config/DbConfig.java
@@ -1,12 +1,0 @@
-package syncMonitor.config;
-
-public interface DbConfig {
-    String getIp();
-    String getId();
-    String getDbLink();
-    String getDbSid();
-    String getPwd();
-    String getPort();
-    String getUser1();
-    String getUser2();
-}
