+++ src/syncMonitor/config/wrapper/DbConfig/DbConfig.java
@@ -1,0 +1,12 @@
+package syncMonitor.config.wrapper.DbConfig;
+
+public interface DbConfig {
+    String getIp();
+    String getId();
+    String getDbLink();
+    String getDbSid();
+    String getPwd();
+    String getPort();
+    String getUser1();
+    String getUser2();
+}
