+++ src/syncMonitor/config/wrapper/MonitorConfig.java
@@ -1,0 +1,12 @@
+package syncMonitor.config.wrapper;
+
+public class MonitorConfig {
+    private String mode;
+    public String getMode() {
+        return mode;
+    }
+
+    public void setMode(String mode) {
+        this.mode = mode;
+    }
+}
