--- src/syncMonitor/dbLink/Dblink.java
@@ -1,8 +1,0 @@
-package syncMonitor.dbLink;
-
-public interface Dblink {
-
-    String doGetTsn();
-    String doGetTime();
-
-}
