+++ src/syncMonitor/mode/OnlyOracle.java
@@ -1,0 +1,4 @@
+package syncMonitor.mode;
+
+public class OnlyOracle {
+}
