+++ src/syncMonitor/mode/view/View.java
@@ -1,0 +1,5 @@
+package syncMonitor.mode.view;
+
+public interface View {
+    void doPrint();
+}
