+++ src/syncMonitor/query/Dblink.java
@@ -1,0 +1,8 @@
+package syncMonitor.query;
+
+public interface Dblink {
+
+    String doGetTsn();
+    String doGetTime();
+
+}
